# api/openapi.yaml
openapi: 3.0.3
info:
  title: GameExplorer API
  version: 0.1.0
  description: |
    Single source of truth for the GameExplorer HTTP contract.
    Go handlers (oapi-codegen strict server) and the TypeScript client
    (openapi-typescript) are generated from this file; never edit generated code.

    Uploads are NOT described here: they use the tus resumable upload protocol
    (https://tus.io/protocols/resumable-upload) mounted at /api/uploads/.

    Every operation requires the session cookie unless it declares `security: []`.
  license:
    name: MIT
servers:
  - url: /api
tags:
  - name: system
  - name: auth
  - name: consoles
  - name: metadata
  - name: uploads
security:
  - sessionCookie: []
paths:
  /health:
    get:
      tags: [system]
      operationId: getHealth
      summary: Liveness probe and startup diagnostics
      description: |
        Always answers 200 while the process is up. `status` is `degraded` when
        a check fails (for example the library is not writable), so the UI can
        explain the problem instead of the container crash-looping.
      security: []
      responses:
        '200':
          description: The process is up.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Health'

  /auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Start a session with the shared password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '204':
          description: Session started; the cookie is set.
          headers:
            Set-Cookie:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'

  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: End the session on this device
      security: []
      responses:
        '204':
          description: Session cookie cleared.
          headers:
            Set-Cookie:
              schema:
                type: string

  /auth/session:
    get:
      tags: [auth]
      operationId: getSession
      summary: Current session
      responses:
        '200':
          description: The session is valid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /consoles:
    get:
      tags: [consoles]
      operationId: listConsoles
      summary: Consoles in carousel order
      responses:
        '200':
          description: All consoles.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Console'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /metadata/games:
    get:
      tags: [metadata]
      operationId: searchGames
      summary: Search games on IGDB, optionally limited to one platform
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 100
        - name: platformId
          in: query
          description: IGDB platform id (Console.igdbPlatformId).
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 25
            default: 10
      responses:
        '200':
          description: Matching games, best match first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MetadataGame'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  /metadata/platforms:
    get:
      tags: [metadata]
      operationId: searchPlatforms
      summary: Search platforms on IGDB (used to add consoles)
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 100
      responses:
        '200':
          description: Matching platforms.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MetadataPlatform'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  /jobs:
    get:
      tags: [uploads]
      operationId: listJobs
      summary: Recent upload jobs, newest first
      description: |
        A job is created when the browser starts a tus upload at /api/uploads/
        (metadata `filename`, optional `consoleSlug`) and follows the file
        through extraction, review and commit.
      responses:
        '200':
          description: Up to 100 jobs.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UploadJob'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /jobs/events:
    get:
      tags: [uploads]
      operationId: streamJobEvents
      summary: Live job changes (server-sent events)
      description: |
        `text/event-stream`. Each change is sent as `event: job` with an
        UploadJob JSON in `data`. On connect the server first replays every
        non-terminal job, then streams changes; a comment line is sent every
        25 seconds to keep proxies from closing the connection.
      responses:
        '200':
          description: Event stream.
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'

  /jobs/{id}:
    get:
      tags: [uploads]
      operationId: getJob
      summary: One upload job
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: The job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadJob'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /jobs/{id}/cancel:
    post:
      tags: [uploads]
      operationId: cancelJob
      summary: Cancel a job and delete its uploaded data
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: The cancelled job.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadJob'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

  /images/{size}/{imageId}:
    get:
      tags: [metadata]
      operationId: getImage
      summary: IGDB image (cover or logo) served from the local cache
      description: |
        Downloaded from the IGDB image CDN on first request and cached under
        DATA_PATH, so the NAS keeps working offline and the browser never
        contacts third parties.
      parameters:
        - name: size
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/ImageSize'
        - name: imageId
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-z0-9]{1,40}$'
      responses:
        '200':
          description: The image.
          headers:
            Cache-Control:
              schema:
                type: string
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/BadGateway'

components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ge_session

  parameters:
    JobId:
      name: id
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-zA-Z0-9_-]{1,64}$'

  responses:
    Conflict:
      description: The resource is in a state that does not allow this operation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Missing, invalid or expired session, or wrong password.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    TooManyRequests:
      description: Too many login attempts from this client; retry later.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    BadRequest:
      description: Invalid parameters.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: The resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    BadGateway:
      description: IGDB failed or could not be reached.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ServiceUnavailable:
      description: IGDB credentials are not configured.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'

  schemas:
    Health:
      type: object
      required: [status, checks]
      properties:
        status:
          type: string
          enum: [ok, degraded]
        checks:
          type: array
          items:
            $ref: '#/components/schemas/HealthCheck'

    HealthCheck:
      type: object
      required: [name, ok]
      properties:
        name:
          type: string
          example: library-writable
        ok:
          type: boolean
        message:
          type: string
          description: Human readable explanation when the check fails.

    LoginRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 1
          format: password

    Session:
      type: object
      required: [expiresAt]
      properties:
        expiresAt:
          type: string
          format: date-time

    Console:
      type: object
      required: [id, slug, displayName, extensions, sortOrder]
      properties:
        id:
          type: integer
          format: int64
        slug:
          type: string
          description: Folder name in the library (ES-DE style).
          example: ps2
        displayName:
          type: string
          example: PlayStation 2
        igdbPlatformId:
          type: integer
          format: int64
          nullable: true
        releaseYear:
          type: integer
          nullable: true
          example: 2000
        logoImageId:
          type: string
          nullable: true
          description: IGDB image id; render with /api/images/logo_med/{logoImageId}.
        extensions:
          type: array
          items:
            type: string
            example: .iso
        sortOrder:
          type: integer

    MetadataGame:
      type: object
      required: [id, name, genres, platformIds]
      properties:
        id:
          type: integer
          format: int64
          description: IGDB game id.
        name:
          type: string
          example: Mario Kart 8 Deluxe
        releaseYear:
          type: integer
          nullable: true
        coverImageId:
          type: string
          nullable: true
          description: Render with /api/images/cover_big/{coverImageId}.
        summary:
          type: string
          nullable: true
        genres:
          type: array
          items:
            type: string
        platformIds:
          type: array
          items:
            type: integer
            format: int64

    MetadataPlatform:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
          format: int64
          description: IGDB platform id.
        name:
          type: string
          example: Wii U
        abbreviation:
          type: string
          nullable: true
        logoImageId:
          type: string
          nullable: true
        releaseYear:
          type: integer
          nullable: true

    ImageSize:
      type: string
      description: IGDB image size preset.
      enum: [cover_small, cover_big, logo_med, screenshot_med]

    JobStatus:
      type: string
      description: |
        uploading → uploaded → (extracting ⇄ needs_password) → review →
        committing → done; failed and cancelled are terminal.
      enum:
        [
          uploading,
          uploaded,
          extracting,
          needs_password,
          review,
          committing,
          done,
          failed,
          cancelled,
        ]

    UploadJob:
      type: object
      required: [id, fileName, size, received, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          description: Equals the tus upload id.
        fileName:
          type: string
          example: INSIDE-Switch-NSP-Base-Game.rar
        size:
          type: integer
          format: int64
        received:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/JobStatus'
        error:
          type: string
          nullable: true
          description: Human readable reason when status is failed.
        originConsole:
          type: string
          nullable: true
          description: Slug of the console screen the upload started from.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ItemKind:
      type: string
      description: What a stored item is within a game folder.
      enum: [base, update, dlc, disc]

    Problem:
      type: object
      description: RFC 9457 problem details.
      required: [title, status]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string