# apps/api/internal/ingestion/infrastructure/sqlite/sqlcgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package sqlcgen

import (
	"database/sql"
)

type Console struct {
	ID             int64
	Slug           string
	DisplayName    string
	IgdbPlatformID sql.NullInt64
	ReleaseYear    sql.NullInt64
	LogoImageID    sql.NullString
	Extensions     string
	DetectorKey    sql.NullString
	SortOrder      int64
	CreatedAt      string
}

type UploadJob struct {
	ID            string
	FileName      string
	Size          int64
	Received      int64
	Status        string
	Error         string
	OriginConsole sql.NullString
	StoragePath   string
	CreatedAt     string
	UpdatedAt     string
}